context/operations/pa/requirements.txt
anthropic>=0.40.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
requests>=2.32.0
